This speeds up things quite a bit when doing lots of small insert queries against Cassandra, by reducing the load on the garbage collector. I'm not 100% sure about the correctness of "no need to include keyspace in routingKeyInfoCache key", but since the session keyspace seems to be immutable, including it in a map key when the map is per-session doesn't seem needed.

benchcmp (when running with benchtime 30s to let some time for the GC to kick in):
```
benchmark                   old ns/op     new ns/op     delta
BenchmarkConnRoutingKey     654           421           -35.63%

benchmark                   old allocs     new allocs     delta
BenchmarkConnRoutingKey     5              4              -20.00%

benchmark                   old bytes     new bytes     delta
BenchmarkConnRoutingKey     112           48            -57.14%
```