@abustany 2015-08-25
Looks like the CI failed due to some unrelated error... Also, this code won't work with Go 1.3, it uses sync.Pool which appeared in 1.4. As 1.5 has just been released, do we want to keep compatibility with Go 1.3?