@abustany 2015-08-25
Another side note: the next "offenders" when it comes to CPU cost when doing lots of small queries are:

- `Session.routingKeyInfo` that uses a lru cache, that lru cache uses a map[interface] as a cache, which is much slower than using a map[string] (we use the statement as a key). Forking that LRU and adapting it to use strings as keys could be an option.
- CQL marshaling. I haven't looked yet deeply into that issue, but since it's based on reflection we probably can't make it much faster.

There also seems to be quite some load on the `select` statements, but I don't see a trivial way to reduce it either.