@abustany 2015-08-25
Adding one more commit, also up for discussion, that uses an LRU cache that does not wrap keys in `interface{}`. Not sure how feasible it is to include an Apache licensed file in an MIT project?

benchcmp:
```
benchmark                   old ns/op     new ns/op     delta
BenchmarkConnRoutingKey     421           335           -20.43%

benchmark                   old allocs     new allocs     delta
BenchmarkConnRoutingKey     4              3              -25.00%

benchmark                   old bytes     new bytes     delta
BenchmarkConnRoutingKey     48            32            -33.33%
```