@Zariel 2015-08-25
Thanks for submitting this, I will take a look. IANAL but I think its OK to include the source with the license, though I would prefer to have it go gettable, is the source available anywhere elese for the lru library? Could you split this PR into 3 parts? One for the changes to the routing key, one for the use statement improvement and finally one for the LRU change.

The test failures look to be caused by Github timing out due to DDoS.