@abustany 2015-08-26
The LRU source is a copy of the original one (from github.com/golang/groupcache/lru), with `interface{}` replaced by `string` for keys. Since I just copied the folder from another repository, it should still be go gettable? ie. `go get github.com/gocql/gocql/lru` once the commit is merged.

I'll split that PR as requested, that's no problem :-)