@abustany 2015-08-26
OK, looks like Travis cached the tests results, so we're still with the spurious errors from last run (I just removed the last two commits that ended up in two separate PRs). Not sure if you can retrigger it? Else, I can rebase it if some other commits get merged...