@abustany 2015-08-27
Here is a benchcmp before and after the commit (before = include keyspace, after = without the keyspace):
```
benchmark                     old ns/op     new ns/op     delta
BenchmarkConnRoutingKey-4     537           326           -39.29%

benchmark                     old allocs     new allocs     delta
BenchmarkConnRoutingKey-4     4              3              -25.00%

benchmark                     old bytes     new bytes     delta
BenchmarkConnRoutingKey-4     112           48            -57.14%
```

So clearly doing a string concatenation every time has a cost :-) If we want to have a global cache, we should maybe use a "map[keyspace]LRU"?