@Zariel 2015-08-31
After some though this is fine, thanks very much!